# docs/uv-operations/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Royal Regent UV Operations",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/uv-operations/machines": {
      "post": {
        "tags": [
          "uv-operations"
        ],
        "summary": "Uv Post Machines",
        "operationId": "uv_post_machines_api_uv_operations_machines_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "uv-operations"
        ],
        "summary": "Uv List Machines",
        "operationId": "uv_list_machines_api_uv_operations_machines_get",
        "parameters": [
          {
            "name": "factory_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Factory Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/internal/uv-agent/enroll": {
      "post": {
        "tags": [
          "uv-agent"
        ],
        "summary": "Enroll",
        "operationId": "enroll_api_internal_uv_agent_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Enroll"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/uv-agent/events/batch": {
      "post": {
        "tags": [
          "uv-agent"
        ],
        "summary": "Events",
        "operationId": "events_api_internal_uv_agent_events_batch_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/uv-agent/heartbeat": {
      "post": {
        "tags": [
          "uv-agent"
        ],
        "summary": "Heartbeat",
        "operationId": "heartbeat_api_internal_uv_agent_heartbeat_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentBatch": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "batch_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Batch Id"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "batch_id",
          "events"
        ],
        "title": "AgentBatch"
      },
...
      "Enroll": {
        "properties": {
          "pairing_code": {
            "type": "string",
            "maxLength": 128,
            "minLength": 24,
            "title": "Pairing Code"
          },
          "enrollment_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Enrollment Id"
          },
          "token": {
            "type": "string",
            "maxLength": 128,
            "minLength": 43,
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pairing_code",
          "enrollment_id",
          "token"
        ],
        "title": "Enroll"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MachineCreate": {
        "properties": {
          "factory_id": {
            "type": "string",
            "const": "huakang-a",
            "title": "Factory Id"
          },
          "operation_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Operation Id"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Expected Version",
            "default": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "title": "Reason",
            "default": ""
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "model": {
            "type": "string",
            "maxLength": 128,
            "title": "Model",
            "default": ""
          },
          "width_mm": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string",
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                  }
                ],
                "decimal_places": 6,
                "gt": 0,
                "max_digits": 18
              },
              {
                "type": "null"
              }
            ],
            "title": "Width Mm"
          },
          "height_mm": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string",
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                  }
                ],
                "decimal_places": 6,
                "gt": 0,
                "max_digits": 18
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Mm"
          },
          "ink_family": {
            "type": "string",
            "maxLength": 64,
            "title": "Ink Family",
            "default": "unknown"
          },
          "maintenance": {
            "type": "boolean",
            "title": "Maintenance",
            "default": false
          },
          "capability_evidence": {
            "type": "string",
            "maxLength": 2000,
            "title": "Capability Evidence",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "factory_id",
          "operation_id",
          "code",
          "name"
        ],
        "title": "MachineCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}